Add tokio-rustls to cargo, (minor) progress on structs and super basic tls cert and key opening, example bits that make sense from echo server in, add comments for what I believe I'll need to do next to structs and general architecting to hook things together into an application with state that does useful things
Yay useful things.

I seem to have a tendency of late to not want to commit before a project is 'just so', where that definition changes depending on the goal in mind / what I'm working on, will compile, will work perfectly, etc.
I dislike waiting until it's (rarely) in one of these states, even if that reveals the mess my stuff really is, at least at the start / when I'm getting back into usage of the language with a second project.
Hence the nonfunctional work-in-progress main.rs is.
(and this arbitrarily verbose commit message I doubt anyone but myself will ever read)